Update housing_market_model.nlogo
Added "memory" to the buyer code. Fixed code to address a repeat offer in a counter-offer situation which was less than the original offer which made no sense. Also fixed it so that if a buyer repeats their offer (eg it had to have been previously declined or countered), they exit the market